Initial commit of dwm puppet module
This commit creates a dwm puppet module that allows
for the compilation and installation of the `dwm` window
manager. It allows for the specification of the desired
version and installation path.

This commit does not provide the ability to manage the
`config.h` file in order to customize dwm prior to compilation.